+++ src/media/audio/format/raw/mod.rs
@@ -1,0 +1,21 @@
+// devela::media::audio::format::raw
+//
+//! Headerless raw PCM audio.
+//
+
+#[cfg(test)]
+mod tests;
+
+mod buf; // PcmRawBuf
+mod error; // PcmRawError
+mod namespace; // PcmRaw
+
+crate::structural_mods! { // _mods
+    _mods {
+        pub use super::{
+            buf::*,
+            error::*,
+            namespace::*,
+        };
+    }
+}