+++ docs/changelog.md
@@ -173,7 +173,7 @@
 ### media::audio
 - new submodules: `acoustic`, `effect`, `format`, `instrument`, `music`, `rhythm`, `synth`.
 - new trait: `PcmSampleType`.
-- new types: `PcmBuffer`, `PcmPlanar`, `PcmSample`, `PcmSpec`, `PcmWav`, `PcmWavBuf`, `PcmWavError`, `PcmWavFmt`.
+- new types: `PcmBuffer`, `PcmPlanar`, `PcmRaw`, `PcmRawBuf`, `PcmRawError`, `PcmSample`, `PcmSpec`, `PcmWav`, `PcmWavBuf`, `PcmWavError`, `PcmWavFmt`.
 
 #### media::visual::color
 - new type `ColorDepth`.